radi timeout napokon

+++ Core/Lib/tactics/tact_1.c
@@ -10,7 +10,7 @@
 
 #include "main.h"
 
-extern int16_t tact_fsm_case;
+extern int32_t tact_fsm_case;
 extern uint32_t tact_delay_1;
 extern uint8_t points;
 extern int8_t tact_state;
@@ -53,10 +53,10 @@ tact_1 ()
 			tact_fsm_case = 40;
 			break;
 		case 40: //bunt pored startnog polja
-			if (get_tact_num_ptr()->side)	// plava
-			cur_task = move_on_path (x_side (-680), -400, phi_side (0), BACKWARD, 0, 0.5, 0, NO_SENS);
+			if (get_tact_num_ptr ()->side)	// plava
+				cur_task = move_on_path (x_side (-680), -400, phi_side (0), BACKWARD, 0, 0.5, 0, NO_SENS);
 			else
-			cur_task = move_on_path (x_side (-670), -400, phi_side (0), BACKWARD, 0, 0.5, 0, NO_SENS);
+				cur_task = move_on_path (x_side (-670), -400, phi_side (0), BACKWARD, 0, 0.5, 0, NO_SENS);
 
 			if (cur_task == TASK_SUCCESS)
 				tact_fsm_case = 45;
@@ -79,10 +79,10 @@ tact_1 ()
 			tact_fsm_case = 70;
 			break;
 		case 70:
-		if (get_tact_num_ptr()->side)	// plava
-			cur_task = task_banner (TO_CENTER);
-		else
-			cur_task = task_banner (TO_OUT);
+			if (get_tact_num_ptr ()->side)	// plava
+				cur_task = task_banner (TO_CENTER);
+			else
+				cur_task = task_banner (TO_OUT);
 
 			if (cur_task == TASK_SUCCESS)
 				tact_fsm_case = 80;
@@ -98,113 +98,122 @@ tact_1 ()
 				tact_fsm_case = 90;
 			break;
 		case 90: //bunt kod protivnickog polja
-		if (get_tact_num_ptr()->side)	// plava
-			cur_task = move_to_xy (x_side (-1182), -597, BACKWARD, V_MAX_DEF,
-						W_MAX_DEF,NO_SENS);
-		else
-			cur_task = move_to_xy (x_side (-1182), -582, FORWARD, V_MAX_DEF,
-						W_MAX_DEF,NO_SENS);
+			if (get_tact_num_ptr ()->side)	// plava
+				cur_task = move_to_xy (x_side (-1182), -597, BACKWARD, V_MAX_DEF,
+				W_MAX_DEF,
+																NO_SENS);
+			else
+				cur_task = move_to_xy (x_side (-1182), -582, FORWARD, V_MAX_DEF,
+				W_MAX_DEF,
+																NO_SENS);
 
 			if (cur_task == TASK_SUCCESS)
-			tact_fsm_case = 100;
-		break;
-	case 100:
-		if (get_tact_num_ptr()->side)	// plava
-			cur_task = move_to_xy(x_side(-1300), -595, BACKWARD, 0.2,
-			W_MAX_DEF, NO_SENS);
-		else
-			cur_task = move_to_xy(x_side(-1300), -580, FORWARD, 0.2,
-			W_MAX_DEF, NO_SENS);
-		if (cur_task == TASK_SUCCESS)
-			tact_fsm_case = 110;
-		break;
+				tact_fsm_case = 100;
+			break;
+		case 100:
+			if (get_tact_num_ptr ()->side)	// plava
+				cur_task = move_to_xy (x_side (-1300), -595, BACKWARD, 0.2,
+				W_MAX_DEF,
+																NO_SENS);
+			else
+				cur_task = move_to_xy (x_side (-1300), -580, FORWARD, 0.2,
+				W_MAX_DEF,
+																NO_SENS);
+			if (cur_task == TASK_SUCCESS)
+				tact_fsm_case = 110;
+			break;
 		case 110:
-		if (get_tact_num_ptr()->side) {
-			grtl_back_grip_all();
-			ruc_back_carry();
-			lift_back_carry();
-			vacuum_back(1);
-			prepare_front();
-		} else {
-			grtl_front_grip_all();
-			ruc_front_carry();
-			lift_front_carry();
-			vacuum_front(1);
-			prepare_back(); }
-		tact_fsm_case = 120;
+			if (get_tact_num_ptr ()->side)
+				{
+					grtl_back_grip_all ();
+					ruc_back_carry ();
+					lift_back_carry ();
+					vacuum_back (1);
+					prepare_front ();
+				}
+			else
+				{
+					grtl_front_grip_all ();
+					ruc_front_carry ();
+					lift_front_carry ();
+					vacuum_front (1);
+					prepare_back ();
+				}
+			tact_fsm_case = 120;
 			break;
 
-	case 120: //rezervisan bunt
-		if (get_tact_num_ptr()->side)	// plava
-			cur_task = move_on_path(x_side(-735), 400, phi_side(90), FORWARD, 0,
-					0.5, 0, NO_SENS);
-		else
-			cur_task = move_on_path(x_side(-685), 400, phi_side(-90), BACKWARD,
-					0, 0.5, 0, NO_SENS);
+		case 120: //rezervisan bunt
+			if (get_tact_num_ptr ()->side)	// plava
+				cur_task = move_on_path (x_side (-735), 400, phi_side (90), FORWARD, 0, 0.5, 0, NO_SENS);
+			else
+				cur_task = move_on_path (x_side (-685), 400, phi_side (-90), BACKWARD, 0, 0.5, 0, NO_SENS);
 
 			if (cur_task == TASK_SUCCESS)
 				tact_fsm_case = 130;
 			break;
 
 		case 130:
-		if (get_tact_num_ptr()->side)	// plava
-			cur_task = rot_to_phi(phi_side(90), W_MAX_DEF, NO_SENS);
-		else
-			cur_task = rot_to_phi(phi_side(-90), W_MAX_DEF, NO_SENS);
+			if (get_tact_num_ptr ()->side)	// plava
+				cur_task = rot_to_phi (phi_side (90), W_MAX_DEF, NO_SENS);
+			else
+				cur_task = rot_to_phi (phi_side (-90), W_MAX_DEF, NO_SENS);
 
-		if (cur_task == TASK_SUCCESS)
+			if (cur_task == TASK_SUCCESS)
 				tact_fsm_case = 140;
 			break;
 
 		case 140:
-		if (get_tact_num_ptr()->side)	// plava
-			cur_task = move_to_xy(x_side(-735), 620, FORWARD, 0.3, W_MAX_DEF,
-					NO_SENS);
-		else
-			cur_task = move_to_xy(x_side(-685), 620, BACKWARD, 0.3, W_MAX_DEF,
-			NO_SENS);
-		if (cur_task == TASK_SUCCESS)
+			if (get_tact_num_ptr ()->side)	// plava
+				cur_task = move_to_xy (x_side (-735), 620, FORWARD, 0.3, W_MAX_DEF,
+				NO_SENS);
+			else
+				cur_task = move_to_xy (x_side (-685), 620, BACKWARD, 0.3, W_MAX_DEF,
+				NO_SENS);
+			if (cur_task == TASK_SUCCESS)
 				tact_fsm_case = 150;
 			break;
 
 		case 150:
-		if (get_tact_num_ptr()->side) {
-			grtl_front_grip_all();
-			ruc_front_carry();
-			lift_front_carry();
-			vacuum_front(1);
-		} else {
-			grtl_back_grip_all();
-			ruc_back_carry();
-			lift_back_carry();
-			vacuum_back(1);
-		}
-		tact_fsm_case = 160;
+			if (get_tact_num_ptr ()->side)
+				{
+					grtl_front_grip_all ();
+					ruc_front_carry ();
+					lift_front_carry ();
+					vacuum_front (1);
+				}
+			else
+				{
+					grtl_back_grip_all ();
+					ruc_back_carry ();
+					lift_back_carry ();
+					vacuum_back (1);
+				}
+			tact_fsm_case = 160;
 			break;
 		case 160: //krece ka polju za gradnju
-		if (get_tact_num_ptr()->side)
-			cur_task = move_to_xy(x_side(-735), -700, BACKWARD, 0.5, W_MAX_DEF,
-			NO_SENS);
-		else
-			cur_task = move_to_xy(x_side(-685), -700, FORWARD, 0.5, W_MAX_DEF,
-			NO_SENS);
+			if (get_tact_num_ptr ()->side)
+				cur_task = move_to_xy (x_side (-735), -700, BACKWARD, 0.5, W_MAX_DEF,
+				NO_SENS);
+			else
+				cur_task = move_to_xy (x_side (-685), -700, FORWARD, 0.5, W_MAX_DEF,
+				NO_SENS);
 			if (cur_task == TASK_SUCCESS)
 				tact_fsm_case = 170;
 			break;
 		case 170:
-		if (get_tact_num_ptr()->side)
-			cur_task = rot_to_phi(phi_side(180), W_MAX_DEF, NO_SENS);
-		else
-			cur_task = rot_to_phi(phi_side(0), W_MAX_DEF, NO_SENS);
+			if (get_tact_num_ptr ()->side)
+				cur_task = rot_to_phi (phi_side (180), W_MAX_DEF, NO_SENS);
+			else
+				cur_task = rot_to_phi (phi_side (0), W_MAX_DEF, NO_SENS);
 			if (cur_task == TASK_SUCCESS)
 				tact_fsm_case = 175;
 			break;
 		case 175:
-			if (get_tact_num_ptr()->side)
-			cur_task = move_on_dir(60, BACKWARD, 0.3, NO_SENS);
-		else
-			cur_task = move_on_dir(60, FORWARD, 0.3, NO_SENS);
-		if (cur_task == TASK_SUCCESS)
+			if (get_tact_num_ptr ()->side)
+				cur_task = move_on_dir (60, BACKWARD, 0.3, NO_SENS);
+			else
+				cur_task = move_on_dir (60, FORWARD, 0.3, NO_SENS);
+			if (cur_task == TASK_SUCCESS)
 				tact_fsm_case = 180;
 			break;
 		case 180:
@@ -224,32 +233,32 @@ tact_1 ()
 				tact_fsm_case = 195;
 			break;
 		case 195:
-		if (get_tact_num_ptr()->side)	// plava
-			cur_task = rot_to_phi(phi_side(0), W_MAX_DEF, NO_SENS);
-		else
-			cur_task = rot_to_phi(phi_side(180), W_MAX_DEF, NO_SENS);
-		if (cur_task == TASK_SUCCESS)
-			tact_fsm_case = 200;
-		break;
+			if (get_tact_num_ptr ()->side)	// plava
+				cur_task = rot_to_phi (phi_side (0), W_MAX_DEF, NO_SENS);
+			else
+				cur_task = rot_to_phi (phi_side (180), W_MAX_DEF, NO_SENS);
+			if (cur_task == TASK_SUCCESS)
+				tact_fsm_case = 200;
+			break;
 
 		case 200:
 			if (get_tact_num_ptr ()->side)	// plava
 				cur_task = task_sprat_12 (FORWARD);
 			else
 				cur_task = task_sprat_12 (BACKWARD);
-		if (cur_task == TASK_SUCCESS)
-			tact_fsm_case = 210;
-		break;
-	case 210: //zavrsetak
-		if (get_tact_num_ptr()->side)
-			cur_task = move_to_xy(x_side(-1000), 500, BACKWARD, 0.5, W_MAX_DEF,
-			NO_SENS);
-		else
-			cur_task = move_to_xy(x_side(-1000), 500, FORWARD, 0.5, W_MAX_DEF,
-			NO_SENS);
-		if (cur_task == TASK_SUCCESS)
-			tact_fsm_case = 9999;
-		break;
+			if (cur_task == TASK_SUCCESS)
+				tact_fsm_case = 210;
+			break;
+		case 210: //zavrsetak
+			if (get_tact_num_ptr ()->side)
+				cur_task = move_to_xy (x_side (-1000), 500, BACKWARD, 0.5, W_MAX_DEF,
+				NO_SENS);
+			else
+				cur_task = move_to_xy (x_side (-1000), 500, FORWARD, 0.5, W_MAX_DEF,
+				NO_SENS);
+			if (cur_task == TASK_SUCCESS)
+				tact_fsm_case = 9999;
+			break;
 
 		}
 	return tact_state;
